requirements.txt
altgraph==0.17.3
Babel==2.12.1
CacheControl==0.13.1
cachetools==5.3.1
certifi==2021.5.30
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==41.0.1
customtkinter==5.1.3
cx-Freeze==6.14.7
cx-Logging==3.1.0
darkdetect==0.8.0
DateTime==5.1
docopt==0.6.2
firebase-admin==6.1.0
gcloud==0.17.0
google-api-core==2.11.1
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
httplib2==0.22.0
idna==2.10
Jinja2==3.0.1
jws==0.1.3
lief==0.12.3
MarkupSafe==2.0.1
msgpack==1.0.5
mysql-connector-python==8.0.32
numpy==1.24.2
oauth2client==3.0.0
pandas==1.5.3
pefile==2023.2.7
Pillow==9.4.0
pipreqs==0.4.13
proto-plus==1.22.2
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.4.3
pygame==2.3.0
pyinstaller==5.12.0
pyinstaller-hooks-contrib==2023.3
PyJWT==2.7.0
pyparsing==3.0.9
Pyrebase==3.0.27
python-dateutil==2.8.2
python-jwt==2.0.1
pytz==2023.3
pywin32-ctypes==0.2.0
requests==2.31.0
requests-toolbelt==0.7.0
rsa==4.9
six==1.16.0
tkcalendar==1.6.1
tqdm==4.65.0
ttkthemes==3.2.2
uritemplate==4.1.1
urllib3==1.26.6
yarg==0.1.9
zope.interface==6.0
